feat: Initialize core application structure, add property details and booking flow, and integrate authentication forms with new components and API services.
