+++ src/components/booking/PaymentPendingView.tsx
@@ -1,0 +1,135 @@
+import React from 'react';
+import { Icon } from "@iconify/react";
+import PageLayout from '../pagelayout/index';
+import usePageTitle from '../../hooks/usePageTitle';
+import { useNavigate } from 'react-router-dom';
+
+interface PaymentPendingViewProps {
+    booking: any;
+    paymentMethod: string;
+    formatPrice: (price: number) => string;
+}
+
+const PaymentPendingView: React.FC<PaymentPendingViewProps> = ({
+    booking,
+    paymentMethod,
+    formatPrice,
+}) => {
+    const navigate = useNavigate();
+    const titleComponent = usePageTitle({
+        title: 'Payment Pending'
+    });
+
+    return (
+        <PageLayout>
+            {titleComponent}
+            <div className="w-full flex flex-col items-center justify-center p-7 mt-20">
+                <div className="lg:w-2/3">
+                    <div className="flex items-center mb-4 visibility:hidden">
+                        <div className="mr-4 cursor-pointer" onClick={() => navigate("/")}>
+                            <svg
+                                width="40"
+                                height="40"
+                                viewBox="0 0 60 60"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <circle cx="30" cy="30" r="30" fill="#191919" />
+                                <path
+                                    d="M32.1377 24.1294L27.139 29.1281C26.5487 29.7184 26.5487 30.6844 27.139 31.2748L32.1377 36.2734"
+                                    stroke="white"
+                                    strokeWidth="2.33538"
+                                    strokeMiterlimit="10"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                />
+                            </svg>
+                        </div>
+                        <h1 className="text-2xl font-medium ml-0">Payment Pending</h1>
+                    </div>
+
+                    <div className="flex flex-col items-center justify-center p-6 printable-section">
+                        <div className="text-center">
+                            {/* Pending Icon and Title */}
+                            <Icon icon="token-branded:wait" className="w-24 h-24 mx-auto mb-1 text-yellow-500" />
+                            <h1 className="text-[22px] font-medium text-gray-800">
+                                Payment Pending!
+                            </h1>
+                        </div>
+                        {/* Amount Section */}
+                        <div className="mt-4 text-center">
+                            <p className="text-[12px] text-gray-600">Amount</p>
+                            <h2 className="text-[20px] font-medium">{formatPrice(booking?.total_charging_fee ?? 0)}</h2>
+                        </div>
+
+                        {/* Booking Details */}
+                        <div className="mt-6 w-full max-w-xl sm:w-full border rounded-lg bg-white shadow-md">
+                            <h3 className="text-md font-semibold text-black px-4 py-3">
+                                Booking Details
+                            </h3>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex justify-between items-center mb-4 px-4 space-x-14">
+                                <p className="text-black font-medium text-[13px]">{formatPrice(booking?.base_price ?? 0)} x {booking?.nights} nights x {booking?.unit_count} unit{booking?.unit_count !== 1 ? "s" : ""}</p>
+                                <p className="text-gray-500 text-[13px]">Total(NGN) {formatPrice(booking?.total_charging_fee ?? 0)}</p>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex justify-between items-center mb-4 px-4">
+                                <p className="text-[14px]">Check-in date</p>
+                                <p className="text-gray-500 text-[13px]">{booking?.check_in_date}</p>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex justify-between items-center mb-4 px-4">
+                                <p className="text-[14px]">Check-out date</p>
+                                <p className="text-gray-500 text-[13px]">{booking?.check_out_date}</p>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex flex-col mb-4 px-4">
+                                <div className="flex justify-between items-center">
+                                    <p className="text-[14px]">Guests</p>
+                                    <p className="text-gray-500 text-[13px]">{(booking?.adults ?? 0) > 0 && <span>{booking?.adults} Adults</span>}</p>
+                                </div>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex justify-between items-center mb-4 px-4">
+                                <p className="font-medium text-[14px]">Transaction Details</p>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex justify-between items-center mb-4 px-4">
+                                <p className="text-[14px]">Amount paid</p>
+                                <p className="text-black font-medium text-[13px]">{formatPrice(booking?.total_charging_fee ?? 0)}</p>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+
+                            <div className="flex justify-between items-center mb-4 px-4">
+                                <p className="text-[14px]">Payment Method</p>
+                                <p className="text-black font-medium text-[13px]">{paymentMethod}</p>
+                            </div>
+
+                            <div className="border-t border-solid border-gray-200 w-full mb-4"></div>
+                            <div className="flex justify-between items-center mb-4 px-4">
+                                <p className="text-[14px]">Payment Status</p>
+                                <p className="text-black font-medium text-[13px]">PENDING</p>
+                            </div>
+                        </div>
+
+                    </div>
+                </div>
+            </div>
+        </PageLayout>
+    );
+};
+
+export default PaymentPendingView;
